.github/workflows/main.yml - use this GitHub Action workflow for building
This action from https://github.com/jeffreytse/jekyll-deploy-action

This deploy action needs to be used instead of the usual GitHub pages
pipeline, because that only allows a whitelisted set of Ruby Gems, and
no custom code for plugins. See [1]

As we want to have some custom code for creating embed frames and
suchlike, we need to use this alternative.