src/_includes/*-list.html - align separating border with list size
The border was being omitted over the bullet list bullets - instead let's make
it run from the left hand side of the list to the right.

This is done by adding a wrapper div around the list, and putting the
border on that instead of the ul list. (And renaming another wrapper
div's class to post-list-content)