src/assets/css/style.css - revert small-screen threshold to 576px
As on on the Macbook I've tried, tiling two windows is enough to push
the result below the threshold of 768px, which seems undesirable. (It
goes to about 620px).

The resolution of the Macbook is 1280x687+ (i.e. the browser has access to
this resolution in fullscreen mode, there may be some height
subtracted for headers)